refactor: input file 이름 수정
